+++ view/BO_postView.php
@@ -17,11 +17,11 @@
                 </h3>
                 
                 <p>
-                    <?= nl2br(htmlspecialchars($post['post_content'])) ?>
+                    <?= nl2br($post['post_content']) ?>
                 </p>
                 <p>
                     <em><a class="modifPost" href="index.php?action=BO_updatePost&amp;id=<?= $post['id'] ?>">[Modifier le Chapitre]</a></em>
-                    <em><a class="deletePost" href="index.php?action=BO_deletePost&amp;id=<?= $post['id'] ?>">[Supprimer le Chapitre]</a></em>
+                    <em><a class="deletePost" href="index.php?action=BO_deletePost&amp;id=<?= $post['id'] ?>" onclick="return(confirm('Etes-vous sûr de vouloir supprimer ce Chapitre ?'));">[Supprimer le Chapitre]</a></em>
                     <em><a class="deleteComment" href="index.php?action=BO_allComments&amp;id=<?= $post['id'] ?>">[Commentaires]</a></em>
                     <br />
                 </p>
